Comments on call_home for testing right now
