+++ fluentd/plugins/input/fluent-plugin-jfrog-siem/lib/fluent/plugin/in_jfrog_siem.rb
@@ -16,8 +16,8 @@ require "fluent/plugin/input"
 require "rest-client"
 require "date"
 require "uri"
-require "fluent/plugin/xray.rb"
-require "fluent/plugin/position_file.rb"
+require "fluent/plugin/xray"
+require "fluent/plugin/position_file"
 
 module Fluent
   module Plugin
@@ -85,7 +85,7 @@ module Fluent
 
 
       def run
-        call_home(@jpd_url)
+        # call_home(@jpd_url)
 
         last_created_date = get_last_item_create_date()
 
@@ -94,9 +94,9 @@ module Fluent
         end
         date_since = last_created_date
         puts "Getting queries from #{date_since}"
-        xray = Xray.new(@jpd_url, @username, @apikey, @wait_interval, @batch_size, @pos_file_path)
+        xray = Xray.new(@jpd_url, @username, @apikey, @wait_interval, @batch_size, @pos_file_path, router)
         violations_channel = xray.violations(date_since)
-        violation_details(violations_channel)
+        xray.violation_details(violations_channel)
         sleep 100
       end
 
@@ -129,104 +129,6 @@ module Fluent
       def get_recent_pos_file()
         pos_file = @pos_file_path + "*.siem.pos"
         return Dir.glob(pos_file).sort[-1]
-      end
-
-      def violation_details(violations_channel)
-        violations_channel.each do |v|
-          Concurrent::Promises.future(v) do |v7|
-            pull_violation_details(v['violation_details_url'])
-            PositionFile.new(@pos_file_path).write(v)
-          end
-        end
-      end
-
-      def pull_violation_details(xray_violation_detail_url)
-        begin
-          detailResp=get_xray_violations_detail(xray_violation_detail_url)
-          time = Fluent::Engine.now
-          detailResp_json = data_normalization(detailResp)
-          router.emit(@tag, time, detailResp_json)
-        rescue => e
-          puts "error: #{e}"
-          raise Fluent::ConfigError, "Error pulling violation details url #{xray_violation_detail_url}: #{e}"
-        end
-      end
-
-      def get_xray_violations_detail(xray_violation_detail_url)
-        response = RestClient::Request.new(
-            :method => :get,
-            :url => xray_violation_detail_url,
-            :user => @username,
-            :password => @apikey
-        ).execute do |response, request, result|
-          case response.code
-          when 200
-            return response.to_str
-          else
-            puts "error: #{response.to_json}"
-            raise Fluent::ConfigError, "Cannot reach Artifactory URL to pull Xray SIEM violations."
-          end
-        end
-      end
-
-      def data_normalization(detailResp)
-        detailResp_json = JSON.parse(detailResp)
-        cve = []
-        cvss_v2_list = []
-        cvss_v3_list = []
-        policy_list = []
-        rule_list = []
-        impacted_artifact_url_list = []
-        if detailResp_json.key?('properties')
-          properties = detailResp_json['properties']
-          for index in 0..properties.length-1 do
-            if properties[index].key?('cve')
-              cve.push(properties[index]['cve'])
-            end
-            if properties[index].key?('cvss_v2')
-              cvss_v2_list.push(properties[index]['cvss_v2'])
-            end
-            if properties[index].key?('cvss_v3')
-              cvss_v3_list.push(properties[index]['cvss_v3'])
-            end
-          end
-
-          detailResp_json["cve"] = cve.sort.reverse[0]
-          cvss_v2 = cvss_v2_list.sort.reverse[0]
-          cvss_v3 = cvss_v3_list.sort.reverse[0]
-          if !cvss_v3.nil?
-            cvss = cvss_v3
-          elsif !cvss_v2.nil?
-            cvss = cvss_v2
-          end
-          cvss_score = cvss[0..2]
-          cvss_version = cvss.split(':')[1][0..2]
-          detailResp_json["cvss_score"] = cvss_score
-          detailResp_json["cvss_version"] = cvss_version
-        end
-
-        if detailResp_json.key?('matched_policies')
-          matched_policies = detailResp_json['matched_policies']
-          for index in 0..matched_policies.length-1 do
-            if matched_policies[index].key?('policy')
-              policy_list.push(matched_policies[index]['policy'])
-            end
-            if matched_policies[index].key?('rule')
-              rule_list.push(matched_policies[index]['rule'])
-            end
-          end
-          detailResp_json['policies'] = policy_list
-          detailResp_json['rules'] = rule_list
-        end
-
-        impacted_artifacts = detailResp_json['impacted_artifacts']
-        for impacted_artifact in impacted_artifacts do
-          matchdata = impacted_artifact.match /default\/(?<repo_name>[^\/]*)\/(?<path>.*)/
-          impacted_artifact_url = matchdata['repo_name'] + ":" + matchdata['path'] + " "
-          impacted_artifact_url_list.append(impacted_artifact_url)
-        end
-        detailResp_json['impacted_artifacts_url'] = impacted_artifact_url_list
-        return detailResp_json
       end
 
     end
